board/geralt_scp: IMGSYS: add RED SCP migration
IMGSYS RED SCP migration

BRANCH=none
BUG=b:317163840
TEST=make BOARD=geralt_scp

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4899976
Commit-Queue: Fei Shao <fshao@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>

+++ baseboard/mtscp-rv32i/baseboard.c
@@ -33,7 +33,8 @@ struct mpu_entry mpu_entries[NR_MPU_ENTRIES] = {
 #if defined(CHIP_VARIANT_MT8195) || defined(CHIP_VARIANT_MT8188)
 	{ CONFIG_DRAM_BASE, DRAM_NC_BASE,
 	  MPU_ATTR_C | MPU_ATTR_W | MPU_ATTR_R },
-	{ DRAM_NC_BASE, KERNEL_BASE + KERNEL_SIZE, MPU_ATTR_W | MPU_ATTR_R },
+	{ DRAM_NC_BASE, (unsigned int)KERNEL_BASE + (unsigned int)KERNEL_SIZE,
+	  MPU_ATTR_W | MPU_ATTR_R },
 #else
 	{ 0x10000000, 0x11400000, MPU_ATTR_W | MPU_ATTR_R },
 #endif
